refactor!: Upgrade plugin to 1.21
This is the first step to bringing this plugin back up to par with modern Minecraft versions. This commit specifically includes all necessary changes to make it compatible with MC 1.21. It additionally includes some needed changes to the build scripts to aid in easier testing of the plugin.

This update will be the last version of the plugin to be compatible with Spigot. After this, I plan to perform a fairly substantial rewrite of it to bring the code up to my current standards of quality. Future updates of this plugin will be built specifically targeting Paper.