feat!: Make use of Mojang's Brigadier command system
This reimplements the commands provided by Waystones under Mojang's in-house command system. Paper provides this to us as an add-in library, so we can use this without needing an extra dependency on this project.

This change will likely make it so the plugin is no longer compatible with Spigot, though there may potentially be some workarounds. This is of course not within the purview of this commit.