chore: Minimize plugin shadowing
This removes Koin and Arrow from the list of shadowed dependencies, and instead sandboxes them using Paper's libraries feature. The Kotlin standard library is also relocated to the plugin namespace to minimize the risk of secondary plugin conflicts.

While this approach is not ideal, it is the best option we have until better tooling exists for testing plugins with dependencies on other plugins.