chore: Add temporary migration command for waystone names
This command is being introduced to allow for the end user to determine when they want to time their database migration. As the plugin provides an SQLite configuration out of the box, we should give server managers time to adjust the config should they decide they want to opt for a MySQL server connection instead.