fix: Name overwrite issue on key linking
We add waystones to the database upon key linking. To ensure we do not accidentally overwrite the names of a waystone which may already exist at a given location, this commit introduces a database check to make sure an entry doesn't already exist for the linked location. This will save us from running the save operation each time we link a new key.