fix: Update parsing of UUID values for MySQL
MySQL seems to handle UUIDs as binary data rather than serialized strings. This led to some weird silent database errors at runtime which looked like MySQL was ignoring queries.

This PR introduces proper serialization of UUIDs in, and additionally provides a bit of clean-up to the DatabaseManager. Additional error logging is included as part of the query functions to ensure issues are not silently ignored in the future.