chore: Improve migration function safety
This ensures we do not run into any runtime errors when attempting to migrate potentially bad data in the legacy warpnames.json file. Additionally, proper logging is now provided to ensure that any failing migrations are logged, and the executor is alerted to the issue upon completion. If the executor doesn't care, they can still force deletion of the old data, but it is not recommended unless they are certain they do not want the data anymore.